Adding basic importing of buckets, changing site title to 'IRIS DynaFed'
Basic <form> being added to index.html where user inputs group name, bucket name, access key and secret key of bucket to import. Currently, user must type the group name, which leaves it open to typos or assigning buckets to groups they have no control over. A future commit hopes to address this by using a drop-down list populated by the list of groups the user is a member of instead. The data is sent via a POST request to a Python script living in cgi-bin which starts the bucket import process and reports back a success or failure to the user as a popup.

Changing title of webpage to 'IRIS DynaFed' to replace generic 'DPMbox Demo' and 'Disk Pool Manager' titles