Fixing import popup not always displaying and bug that logged bucket access and secret keys
The form for importing buckets was sending data to the backend via a GET request, which meant form fields including access and secret key were showing up in Apache logs. Fixing this as it is a big security risk. Form fields are no longer sent and are instead retrieved by the importBucket function, which prevents any accidental logging of sensitive data.

Also fixing an issue where the popup indicating a success or failure of bucket importing wasn't always displaying. The popup now always displays and the page is moved to the top to bring focus back to the buckets.