Bucket import and remove forms now generated in JS, JS fetches user groups to populate group list dropdown
Removing the forms in index.html for importing and removing buckets as they are now generated and populated with information in backend JS (in dpmbox-ui.js). This is because an extra request is made to get a list of the user's groups to populate the group <select> dropdown in the forms.

Dynamically generating import and remove forms in dpmbox-ui.js now. When the webpage finishes loading, a HEAD request is made to DynaFed to get a new header containing a list of the user's IAM groups. Once we have this response, it is sent to a function which generates the forms and populates the group <select> field. These forms are then appended to the <body> in index.html.

This all ensures that the user is restricted in what groups they can assign or remove buckets to, preventing them from managing buckets and groups they have no access to.