Replacing "Unauthorized 401" error with more informative popup message
An "Unauthorized 401" error occurs when the user navigates to another directory after their OIDC session has expired due to inactivity. Previously, the generic HTTP Unauthorized dialogue message would be displayed and it could be hard to figure out what it meant or what to do next. Replacing this with a more simple popup message that refreshes the page when the user hits "Authenticate". This either restarts the same session or redirects to the IAM login page to start a new one.