Removing duplicate 'Modified' filter, fixing bug with incorrect xhr status code being consulted
There are two Modified filters in the bucket contents UI that do virtually the exact same thing when ticked. Commenting one out.

The incorrect status code for one of the XMLHttpRequests being performed on a file upload was being consulted. As such, on the event of a successful upload, the status code for the OPTIONS request was being consulted instead of the PUT request to determine if the upload was a success. Fixing this by creating a new xhr variable, xhr2