Adding bulk downloading of files, removing "double click to download" functionality
This allows for multiple selected file in the same bucket to be downloaded at the same time as part of a zip file. If one file is downloaded, it is not zipped. Asynchronously, one XMLHttpRequest is sent to GET each object through their presigned URLs. When an object is successfully retrieved, it is sent to a function that compiles it as part of a global zip file. The zip file is a global variable so that each spawned XHR can add its contents to it.

Also removing the ability to double click a file to download it. This is because it often resulted in incorrect behaviour, such as index.html being downloaded instead of the requested file. Why this happens is unknown but the double click functionality may be replaced with functionality to display file metadata instead. With this commit, double clicking now does nothing.