Asynchronous file download solution (non-zip)
This replaces the zipped file solution present in development branch for bulk downloading. This allows the user to download files directly from Echo with DynaFed doing no work besides fetching the presigned URLs and performing the relevant GET requests.

Because this solution is more lightweight, there is no limit on the number of files the user can download at once. If more than one file is selected, a confirmation box is displayed asking them to confirm, just to make them aware that they are accepting multiple downloads at once.

The only downside to this solution is how browsers typically handle multiple downloads and how the user would go about cancelling downloads. There is typically no way to cancel all downloads at once in a browser, requiring the user to press cancel on each one. If lots of files are being downloaded, the pace of incoming downloads can overwhelm a user trying to cancel them. If the files are small, this usually means they will be fully downloaded before the user can cancel them.