Improving file downloading, hiding upload button, adding more helpful popup messages when importing/removing, removing, removing duplicate node warning message
This commit aims to rectify problems that arise when browsers are told to download lots of files at once. Instead of telling the browser to download all files at once (which results in the browser randomly cancelling some downloads), downloads are pushed onto a callstack every two seconds through a setTimeout. This allows for the callstack to grow slowly enough that the browser can handle all the downloads it can potentially be making. Also restoring the original file download function as this works best in multiple browsers for downloading single files.

Hiding the upload button on the UI (hence preventing all file upload functionality) as this feature is not ready and just results in errors if the user were to click on it. So there's little point in it being available right now.

DynaFed server has been reconfigured to automatically reload httpd service every 10 mins, meaning buckets can take up to 10 mins to import. The user is now informed of this through a popup on a successful import. Also added extra text to import/remove failure popups to inform users as to why the operation may have failed.

A warning message is output to console every time w2ui tries to update the sidebar but a node (directory/bucket) already exists. Due to its slightly buggy nature, this message is displayed far too often, even when a simple refresh takes place. Removing this warning message as it doesn't result in any misbehaviour or errors and the message is just distracting to the end user.