Adding download support for files that open in browser instead of download
Files with certain extensions, such as .txt or .xml, open in the browser when we attempt to download them from Echo. This is because the files are coming cross-origin and Echo doesn't return a "Content-Disposition: attachment" header that can force a download. So instead what happens is, if the browser can open the file instead of download it, it will. This causes problems with some files being downloaded from Echo

This commit solves this by checking the file extension of the file being downloaded. If it is in a list of file extensions that open instead of download, it downloads the files to the server through an XMLHttpRequest instead and then sends it to the user. This isn't ideal and could be problematic if the files were big but most of the file extensions concerned refer to files that typically aren't that big.