Removing redundant download progress bar, re-authenticating user without informing them
Leftover code from when zipping multiple files during download was the solution is being removed as it is redundant.

Removing a popup that informs the user their session has expired before prompting them to re-authenticate. This happens often enough to annoy the user so the process happens in the background instead (is usually just a page refresh). This happens when the Apache service is reloaded every 10 mins