Adding basic admin functionality, starting with importing and removing any buckets
This still requires the keys to the bucket to be removed but this will be factored out in the next commit. At the moment, a basic button is controlling a cookie's contents for switching the admin functionality on and off. When on, the import bucket popup replaces the group dropdown with a text input, allowing any "group" to be assigned a bucket (theoretically anything could be inserted here so good for setting up DynaFed for a group that doesn't exist yet). Also when admin state on, in the remove bucket popup, the group dropdown contains all groups currently in DynaFed.