Adding dynamic blacklisting functionality for admins
This allows admins to specify a bucket to blacklist and remove an existing blacklisted bucket from the list. This comes in the form of additional popups for admins to enter this required information. They are accessed by two buttons, one for adding to blacklist and one for removing from blacklist, which appear in the list of blacklisted buckets popup.