Blacklist only viewable by admins, custom popup messages depending on error from backend
Decided to make blacklist only viewable by admins. Blacklist button appears and disappears as admin mode is turned on and off. Otherwise, this could alert users to the existence of S3 buckets and make them wonder why they've been blacklisted in the first place (sensitive data inside?).

Adding customised messages for the popup that displays when there's an error for importing or removing buckets. Still to do the same for admins adding to and removing from the blacklist.