Adding more detailed popup messages for blacklist change errors, adding string validation for group and bucket names, adding warning message for admin operations
Popup messages now display more detail if errors occur, such as failure to sync files, bucket already present in list, etc.

Adding additional red warning text on boxes that perform admin operations. This informs the admin that they should take care with the operation as it will impact live users.

Adding group and bucket name validation for importing and removing buckets and adding to and removing from blacklist. A failure to match the regex results in a popup informing the user of an invalid name.