Fixing a buggy error message in the event of a failed file download
If a file fails to download, an extra error message is displayed because 'url' is undefined. This is the url of the downloading file and is undefined because the download failed. The additional error message exposes some code. This commit fixes it by displaying a generic popup error message and not logging code to the console.