Spread out new functionality across new .js files, updated README
This allows for new code that had previously been tacked on the end of dpmbox-ui.js to be separated into distinct JS files relevant to its functionality. It also allows for better maintenance and readability.

Also updated README to reflect the updated project, since it is now quite different from the original forked project.