Users can only see groups they are a member of, admins can see all groups and buckets
This adds some further privacy to groups and their buckets. A user can now only navigate to buckets assigned to groups they are a member of, provided that group has imported buckets. If one of their groups has not imported buckets, the group is not displayed but the user can still import buckets for them. If the user tries to navigate to a group they are not a member of through the address bar, it won't work and nothing will be displayed (except the gridpp folder to navigate them back to the list of groups).

If an admin is signed in and has admin mode on, they can view and access all groups and buckets. Once they turn admin mode off, they lose this access.